refactor dataloaders parameter form
